server/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "WPP Manager 后端"
authors = [
    {name = "hydra", email = "zbhydra@163.com"}
]
dependencies = [
    "aiomysql>=0.3.2",
    "types-PyYAML>=6.0.12.20250915",
    "httpx[socks]>=0.28.1",
    "bcrypt>=4.0.0",
    "fastapi>=0.124.4",
    "uvicorn[standard]>=0.38.0",
    "gunicorn>=23.0.0",
    "sqlalchemy[mypy]>=2.0.45",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.0",
    "email-validator>=2.3.0",
    "redis>=7.1.0",
    "aiosmtplib>=3.0.2",
    "cryptography>=46.0.3",
    "greenlet>=3.3.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[project.optional-dependencies]
# 开发依赖
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "black>=25.12.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "ruff>=0.14.9",
    "httpx>=0.28.1",
    "pre-commit>=4.0.1",
    "bandit[toml]>=1.7.7",
    "types-passlib>=1.7.7.20250602",
    "types-redis>=4.6.0.20241004",
    "types-pyjwt>=1.7.1",
]

[project.scripts]
server = "app.main:main_entry"
dev = "app.main:main_entry --reload"


[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.11",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["src", "tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
check_untyped_defs = false
no_implicit_optional = false
plugins = "sqlalchemy.ext.mypy.plugin"
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true


[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
ignore = ["E402"]  # Ignore module level import not at top of file
per-file-ignores = {"__init__.py" = ["F401"], "tests/*" = ["B011", "S101"]}  # Allow assert and subprocess in tests

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::pytest.PytestUnraisableExceptionWarning",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "build", "dist"]
skips = ["B101", "B601"]  # Skip assert_used and shell_injection_common for this project
